# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Manuskript authoring and Story Oracle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video/projects:
    get:
      operationId: listVideoProjects
      tags: [content-creators]
      summary: List the authenticated user's content-creation projects
      responses:
        "200":
          description: Projects the user owns or is a member of
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VideoProject" }
        "401":
          description: Authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: createVideoProject
      tags: [content-creators]
      summary: Create a locked video project (Captain)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoProjectInput" }
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoProjectDetail" }
        "400":
          description: Invalid project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}:
    get:
      operationId: getVideoProject
      tags: [content-creators]
      summary: Read a permitted video project with members and assets
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoProjectDetail" }
        "403":
          description: Not allowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/members:
    post:
      operationId: addVideoProjectMember
      tags: [content-creators]
      summary: Invite a member by email with a leg role (Captain only)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoMemberInput" }
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoMember" }
        "400":
          description: Invalid member or user not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/assets:
    get:
      operationId: listVideoAssets
      tags: [content-creators]
      summary: List the locked assets in a project vault
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      responses:
        "200":
          description: Vault assets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VideoAsset" }
        "403":
          description: Not allowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: uploadVideoAsset
      tags: [content-creators]
      summary: Upload raw footage into the locked vault
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/VideoAssetUploadInput" }
      responses:
        "201":
          description: Uploaded asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoAsset" }
        "400":
          description: Invalid upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  parameters:
...
    VideoProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VideoProject:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, ownerId, name, description, status, createdAt, updatedAt]
    VideoProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 2000 }
      required: [name]
    VideoProjectDetail:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        myRole: { type: ["string", "null"] }
        members:
          type: array
          items: { $ref: "#/components/schemas/VideoMember" }
        assets:
          type: array
          items: { $ref: "#/components/schemas/VideoAsset" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, ownerId, name, description, status, myRole, members, assets, createdAt, updatedAt]
    VideoMember:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        userId: { type: string }
        role: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, userId, role, status, createdAt]
    VideoMemberInput:
      type: object
      properties:
        email: { type: string, format: email }
        role:
          type: string
          enum: [UPLOADER, ARCHITECT, VISUAL_EDITOR, SOUND_DESIGNER, MOTION_COLOR, VIEWER]
      required: [email, role]
    VideoAsset:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        uploaderId: { type: string }
        kind: { type: string }
        fileName: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        durationMs: { type: ["integer", "null"] }
        status: { type: string }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, uploaderId, kind, fileName, mimeType, sizeBytes, durationMs, status, version, createdAt]
    VideoAssetUploadInput:
      type: object
      description: Multipart upload of a raw asset into the project vault
      properties:
        file:
          type: string
          format: binary
        kind:
          type: string
          enum: [RAW_VIDEO, RAW_AUDIO, SCREEN_REC, B_ROLL, REFERENCE, VO_PICKUP, GRAPHIC]
      required: [file, kind]
...